// src/app/model/Employee.ts
export class Employee{

     empid: number;
     name: string;
     city: string;
     state: string;
     contactNo: string;
     address: string;
     emailid: string;
     pinCode: string;

     constructor(){
        this.address ='';
        this.name ='';
        this.city='';
        this.state='';
        this.contactNo='';
        this.emailid='';
        this.empid=1;
        this.pinCode='';
     }

}